Codechange: Allow using more than 65536 NewGRF string IDs.
NewGRF string allocation allowed up to 524288 strings, however stringid was passed as uint16_t which limits to 2^16.

+++ src/newgrf_text.cpp
@@ -636,7 +636,7 @@ const char *GetGRFStringFromGRFText(const GRFTextWrapper &text)
 /**
  * Get a C-string from a stringid set by a newgrf.
  */
-const char *GetGRFStringPtr(uint16_t stringid)
+const char *GetGRFStringPtr(uint32_t stringid)
 {
 	assert(stringid < _grf_text.size());
 	assert(_grf_text[stringid].grfid != 0);
